Safari-609.1.11.1/LayoutTests/fast/events/onunload.html

<html>
<body onunload="unload()" onload="load()">
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function load()
{
    location = "data:text/html,You should have seen an unload alert appear.<script>if (window.testRunner) testRunner.notifyDone(); </" + "script>";
}

function unload()
{
    alert("unload");
}
</script>
start document
</body>
</html>
