LayoutTests/fast/events/onunload-expected.txt

main frame - has 1 onunload handler(s)
ALERT: unload
You should have seen an unload alert appear.
